Key Value Field

Generic for server generation

Generic for stack/block removal

Necessary to be able to remove the block from the stack.

Necessary to be able to remove the stack from the workspace.

Handle initialisation of workspace, stacks, ...

Usage

# S3 method for keyvalue_field
generate_server(x, ...)

generate_server(x, ...)

# S3 method for block
generate_server(x, ...)

# S3 method for data_block
generate_server(x, id, ...)

# S3 method for upload_block
generate_server(x, id, ...)

# S3 method for transform_block
generate_server(x, in_dat, id, ...)

# S3 method for plot_block
generate_server(x, in_dat, id, ...)

# S3 method for ggiraph_block
generate_server(x, in_dat, id, ...)

# S3 method for stack
generate_server(x, id = NULL, new_block = NULL, ...)

handle_remove(x, ...)

# S3 method for block
handle_remove(x, vals, session = getDefaultReactiveDomain(), ...)

# S3 method for stack
handle_remove(x, vals, session = getDefaultReactiveDomain(), ...)

# S3 method for workspace
generate_server(x, id = NULL, ...)

init(x, ...)

# S3 method for workspace
init(x, stacks, vals, session, ...)

# S3 method for stack
init(x, vals, session, ...)

Arguments

x

Element.

...

Generic consistency.

id

Unique module id. Useful when the workspace is called as a module.

in_dat

Reactive input data

new_block

For dynamically inserted blocks.

vals

Internal reactive values.

session

Shiny session object.

stacks

List of workspace stacks.