Blocks

Blocks consist of a set of fields, an expression that defines what the block produces (given the result of the previous block combined with user input), plus some meta data.

Usage

new_block(fields, expr, name = rand_names(), ..., class = character())

is_block(x)

is_initialized(x)

# S3 method for block
is_initialized(x)

initialize_block(x, ...)

generate_code(x)

# S3 method for block
generate_code(x)

# S3 method for arrange_block
generate_code(x)

# S3 method for transform_block
generate_code(x)

# S3 method for data_block
generate_code(x)

# S3 method for call
generate_code(x)

evaluate_block(x, ...)

# S3 method for data_block
evaluate_block(x, ...)

# S3 method for transform_block
evaluate_block(x, data, ...)

# S3 method for plot_block
evaluate_block(x, data, ...)

# S3 method for plot_layer_block
evaluate_block(x, data, ...)

new_data_block(
  ...,
  dat = as.environment("package:datasets"),
  selected = character()
)

data_block(...)

new_upload_block(...)

upload_block(...)

new_filesbrowser_block(volumes = c(home = path.expand("~")), ...)

filesbrowser_block(...)

new_parser_block(data, expr, fields = list(), ..., class = character())

new_csv_block(data, ...)

csv_block(data, ...)

new_rds_block(data, ...)

rds_block(data, ...)

new_json_block(data, ...)

json_block(data, ...)

new_sas_block(data, ...)

sas_block(data, ...)

new_xpt_block(data, ...)

xpt_block(data, ...)

# S3 method for data_block
initialize_block(x, ...)

new_filter_block(
  data,
  columns = colnames(data)[1L],
  values = character(),
  filter_fun = "==",
  ...
)

filter_block(data, ...)

new_select_block(data, columns = colnames(data)[1], ...)

new_summarize_block(
  data,
  func = c("mean", "se"),
  default_columns = character(),
  ...
)

summarize_block(data, ...)

select_block(data, ...)

new_arrange_block(data, columns = colnames(data)[1], ...)

arrange_block(data, ...)

new_group_by_block(data, columns = colnames(data)[1], ...)

group_by_block(data, ...)

new_join_block(
  data,
  y = data(package = "blockr.data")$result[, "Item"],
  type = character(),
  by_col = character(),
  ...
)

join_block(data, ...)

new_head_block(data, n_rows = numeric(), n_rows_min = 1L, ...)

head_block(data, ...)

# S3 method for transform_block
initialize_block(x, data, ...)

# S3 method for default
initialize_block(x, data, ...)

# S3 method for plot_block
initialize_block(x, data, ...)

update_fields(x, ...)

# S3 method for data_block
update_fields(x, session, ...)

# S3 method for transform_block
update_fields(x, session, data, ...)

# S3 method for plot_block
update_fields(x, session, data, ...)

# S3 method for field
is_initialized(x)

mutate_block(data, ...)

Arguments

fields

A list of field, each entry inheriting from "field"

expr

A quoted expression (compatible with partial substitution as implemented in base::bquote() and intended for evaluation in the context of the fields)

name

Block name

...

Further (metadata) attributes

class

Block subclass

x

An object inheriting form "block"

data

Block input data

dat

Multiple datasets.

selected

Selected dataset.

volumes

Paths accessible by the shinyFiles browser.

columns

Column(s) to select.

values

Definition of the equality filter.

filter_fun

Default filter fun for the expression.

func

Summarize function to apply.

default_columns

If you know in advance each function to apply, you can also pass predefined selected column for each summary. Therefore when not of length 0, columns should have the same length as func.

y

Second data block.

type

Join type.

by_col

If you know in advance which column you want to join

n_rows

Number of rows to return.

n_rows_min

Minimum number of rows.

session

Shiny session

Note

Useful to build other parser blocks