Blocks

Blocks consist of a set of fields, an expression that defines what the block produces (given the result of the previous block combined with user input), plus some meta data.

new_block(fields, expr, name = rand_names(), ..., class = character())

is_block(x)

is_initialized(x)

# S3 method for block
is_initialized(x)

initialize_block(x, ...)

generate_code(x)

# S3 method for block
generate_code(x)

# S3 method for transform_block
generate_code(x)

evalute_block(x, ...)

# S3 method for data_block
evalute_block(x, ...)

# S3 method for transform_block
evalute_block(x, data, ...)

# S3 method for plot_block
evalute_block(x, data, ...)

new_data_block(...)

data_block(...)

# S3 method for data_block
initialize_block(x, ...)

new_filter_block(data, columns = colnames(data)[1L], values = character(), ...)

filter_block(data, ...)

new_select_block(data, columns = colnames(data)[1], ...)

new_summarize_block(
  data,
  column = colnames(data)[1L],
  func = c("mean", "sd"),
  ...
)

summarize_block(data, ...)

select_block(data, ...)

arrange_block(data, ...)

group_by_block(data, ...)

new_plot_block(
  data,
  plot_opts = list(colors = c("blue", "red"), point_size = 3, title = "Plot title", theme
    = c("theme_minimal", "theme_gray", "theme_linedraw", "theme_dark", "theme_light",
    "theme_classic", "theme_void", "theme_bw"), x_lab = "X axis label", y_lab =
    "Y axis label", errors = list(show = FALSE, ymin = character(), ymax = character()),
    lines = list(show = FALSE, group = character(), color = character())),
  ...
)

plot_block(data, ...)

new_cheat_block(data, ...)

cheat_block(data, ...)

# S3 method for transform_block
initialize_block(x, data, ...)

# S3 method for plot_block
initialize_block(x, data, ...)

update_fields(x, ...)

# S3 method for data_block
update_fields(x, session, ...)

# S3 method for transform_block
update_fields(x, session, data, ...)

# S3 method for plot_block
update_fields(x, session, data, ...)

# S3 method for field
is_initialized(x)

Arguments

fields

A list of field, each entry inheriting from "field"

expr

A quoted expression (compatible with partial substitution as implemented in base::bquote() and intended for evaluation in the context of the fields)

name

Block name

...

Any other params. TO DO

class

Block subclass

x

An object inheriting form "block"

data

Block input data

columns

Column(s) to select.

values

Definition of the equality filter.

column

Column to select.

func

Summarize function to apply.

plot_opts

List containing options for ggplot (color, ...).

session

Shiny session