Blocks

Blocks consist of a set of fields, an expression that defines what the block produces (given the result of the previous block combined with user input), plus some meta data.

Usage

new_block(
  fields,
  expr,
  name = rand_names(),
  ...,
  class = character(),
  layout = default_layout_fields
)

is_block(x)

is_initialized(x)

# S3 method for block
is_initialized(x)

initialize_block(x, ...)

generate_code(x)

# S3 method for block
generate_code(x)

# S3 method for transform_block
generate_code(x)

evaluate_block(x, ...)

# S3 method for data_block
evaluate_block(x, ...)

# S3 method for transform_block
evaluate_block(x, data, ...)

# S3 method for plot_block
evaluate_block(x, data, ...)

# S3 method for ggiraph_block
evaluate_block(x, data, ...)

new_data_block(
  ...,
  dat = as.environment("package:datasets"),
  selected = character()
)

# S3 method for plot_block
evaluate_block(x, data, ...)

data_block(...)

# S3 method for data_block
initialize_block(x, ...)

new_filter_block(
  data,
  columns = colnames(data)[1L],
  values = character(),
  filter_fun = "==",
  ...
)

filter_block(data, ...)

new_select_block(data, columns = colnames(data)[1], ...)

new_summarize_block(
  data,
  func = c("mean", "se"),
  default_columns = character(),
  ...
)

summarize_block(data, ...)

select_block(data, ...)

arrange_block(data, ...)

group_by_block(data, ...)

new_join_block(
  data,
  y = data(package = "blockr.data")$result[, "Item"],
  type = character(),
  by_col = character(),
  ...
)

join_block(data, ...)

new_head_block(data, n_rows = numeric(), n_rows_min = 1L, ...)

head_block(data, ...)

new_plot_block(
  data,
  plot_opts = list(colors = c("blue", "red"), point_size = 3, title = "Plot title", theme
    = c("theme_minimal", "theme_gray", "theme_linedraw", "theme_dark", "theme_light",
    "theme_classic", "theme_void", "theme_bw"), x_lab = "X axis label", y_lab =
    "Y axis label", errors = list(show = FALSE, ymin = character(), ymax = character()),
    lines = list(show = FALSE, group = character(), color = character())),
  ...
)

plot_block(data, ...)

new_ggiraph_block(
  data,
  plot_opts = list(colors = c("blue", "red"), point_size = 3, title = "Plot title", theme
    = c("theme_minimal", "theme_gray", "theme_linedraw", "theme_dark", "theme_light",
    "theme_classic", "theme_void", "theme_bw"), x_lab = "X axis label", y_lab =
    "Y axis label", errors = list(show = TRUE, ymin = character(), ymax = character()),
    lines = list(show = TRUE, group = character(), color = character())),
  ...
)

ggiraph_block(data, ...)

# S3 method for transform_block
initialize_block(x, data, ...)

# S3 method for default
initialize_block(x, data, ...)

# S3 method for plot_block
initialize_block(x, data, ...)

update_fields(x, ...)

# S3 method for data_block
update_fields(x, session, ...)

# S3 method for transform_block
update_fields(x, session, data, ...)

# S3 method for plot_block
update_fields(x, session, data, ...)

# S3 method for ggiraph_block
update_fields(x, session, data, ...)

new_cheat_block(data, ...)

cheat_block(data, ...)

new_asfactor_block(data, column = "VISIT", ...)

asfactor_block(data, ...)

demo_data_block(...)

demo_join_block(data, ...)

demo_arrange_block(data, ...)

demo_group_by_block(data, ...)

demo_filter_block_1(data, ...)

demo_filter_block_2(data, ...)

demo_summarize_block(data, ...)

# S3 method for field
is_initialized(x)

new_mutate_block(data, value = NULL, ...)

mutate_block(data, ...)

Arguments

fields

A list of field, each entry inheriting from "field"

expr

A quoted expression (compatible with partial substitution as implemented in base::bquote() and intended for evaluation in the context of the fields)

name

Block name

...

Any other params. TO DO

class

Block subclass

layout

Callback function accepting one argument: the list of fields to layout and returns one or more UI tag(s).

x

An object inheriting form "block"

data

Tabular data in which to select some columns.

dat

Multiple datasets.

selected

Selected dataset.

columns

Column(s) to select.

values

Definition of the equality filter.

filter_fun

Default filter fun for the expression.

func

Summarize function to apply.

default_columns

If you know in advance each function to apply, you can also pass predefined selected column for each summary. Therefore when not of length 0, columns should have the same length as func.

y

Second data block.

type

Join type.

by_col

If you know in advance which column you want to join

n_rows

Number of rows to return.

n_rows_min

Minimum number of rows.

plot_opts

List containing options for ggplot (color, ...).

session

Shiny session

column

Column to apply the operation on.

value

Intial value for muate block